+++ setup.py
@@ -22,6 +22,7 @@ setup(
     packages =          find_packages(),
     python_requires =   ">=3.10",
     install_requires =  [
-
+                            "matplotlib",
+                            "torch"
                         ]
 )
