+++ utilities/arguments/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ["ARGS"]
+
+from utilities.arguments.primary    import ARGS
