--- utils/__init__.py
@@ -1,8 +1,0 @@
-"""Utilities package."""
-
-__all__ = ["arguments", "banner", "logger"]
-
-from utils.arguments    import ARGS
-from utils.banner       import BANNER
-from utils.logger       import LOGGER
-from utils.timestamp    import TIMESTAMP
