--- utils/arguments/__init__.py
@@ -1,3 +1,0 @@
-__all__ = ["primary"]
-
-from utils.arguments.primary    import ARGS
